fixed issues with packages and gulp task not working with new components. This means we can now produce the application.min.js file again
