+++ index.html
@@ -62,16 +62,7 @@
           <p class="lead">A chance for you anyone to come along, get involved and help us shape this project. Venue TBC very soon.</p>
         </section>
       </div>
-    </div> 
-
-<!-- <div class="container">
-  <div class="jumbotron">
-    <section class="text-center">
-      <h2>Public meeting on Saturday 6th February 2016 4pm - 6pm</h2>
-      <p class="lead">A chance for you anyone to come along, get involved and help us shape this project. The meeting will be held at Gulliver's, Oldham st, M4 1LW</p>
-    </section>
-  </div>
-</div> -->
+    </div>
 
   <div class="container">
     <img class="center-block" src="./assets/images/HeaderFactoryLogo.jpg" />
@@ -79,9 +70,9 @@
         <div id="partisan-app" />
     </section>
   </div>
-  <!-- <script src="./scripts/application.min.js"></script> -->
+  <script src="./scripts/application.min.js"></script>
 
-  <script type="text/javascript" src="node_modules/react/dist/react-with-addons.js"></script>
+  <!-- <script type="text/javascript" src="node_modules/react/dist/react-with-addons.js"></script>
   <script type="text/javascript" src="node_modules/react-dom/dist/react-dom.js"></script>
   <script type="text/javascript" src="node_modules/director/build/director.js"></script>
   <script type="text/javascript " src="scripts/utils.js "></script>
@@ -94,7 +85,7 @@
   <script type="text/javascript " src="scripts/components/inspiration.js "></script>
   <script type="text/javascript " src="scripts/components/contact.js "></script>
 
-  <script type="text/javascript " src="scripts/application.js "></script>
+  <script type="text/javascript " src="scripts/application.js "></script> -->
 </body>
 <!-- <footer class="footer">
   <div class="container">
