+++ tsconfig.json
@@ -28,8 +28,7 @@
         "typings/react/react-global.d.ts",
         "typings/react/react.d.ts",
         "typings/tsd.d.ts",
-        "scripts/aboutus.tsx",
-        "scripts/application.tsx",
+        "scripts/app.tsx",
         "scripts/components/cafebar.tsx",
         "scripts/components/contact.tsx",
         "scripts/components/howareweorganised.tsx",
